requirements.txt
langchain-community
langchain-openai
openai
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
websockets
requests
streamlit